CC-35256: Consolidate WireMock product-header transformer and fix leftover-record flakiness in network ITs
Fold ElasticProductHeaderTransformer into ElasticSearchMockUtil as a shared
constant instead of a standalone class, alongside its sibling header helpers.

testTooManyRequests and testServiceUnavailable wrote one more record than
batch.size, relying on master's synchronous BulkProcessor to fail before the
leftover record was ever buffered. The async BulkIngester never blocks on
add(), so the leftover record lands in the buffer regardless, and close()'s
now-correct unconditional drain (bf11247a) gives it its own delivery attempt,
doubling the observed bulk request count. Write exactly batch.size records so
nothing is left to strand.
